PPF-359 - Move assertions to test files instead of inside helpers
