+++ e2e/tests/integrations/admin/create-organization.test.ts
@@ -1,8 +1,11 @@
-import { test } from "@playwright/test";
-import { createOrganization } from "./helpers";
+import { test, expect } from "@playwright/test";
+import { createOrganization } from "./create-organization";
 
 test.use({ storageState: "playwright/.auth/admin.json" });
 
 test("As an admin I can create a new organization", async ({ page }) => {
-  await createOrganization(page);
+  const {page: organizerPage, organizationName} = await createOrganization(page);
+  await expect(
+    organizerPage.locator("h1").getByText(`Organization Details: ${organizationName}`)
+  ).toBeVisible();
 });
